implement get cla api and refactor binding cla api

+++ controllers/org-repo.go
@@ -2,6 +2,7 @@ package controllers
 
 import (
 	"encoding/json"
+	"fmt"
 
 	"github.com/astaxie/beego"
 
@@ -15,24 +16,44 @@ type OrgRepoController struct {
 // @Title Bind CLA to Org/Repo
 // @Description bind cla
 // @Param	body		body 	models.OrgRepo	true		"body for org-repo content"
-// @Success 200 {int} models.OrgRepo
+// @Success 201 {int} models.OrgRepo
 // @Failure 403 body is empty
 // @router / [post]
 func (this *OrgRepoController) Post() {
+	var statusCode = 201
+	var reason error
+
 	defer func() {
-		this.ServeJSON()
+		sendResponse(&this.Controller, statusCode, reason)
 	}()
 
 	var orgRepo models.OrgRepo
 
 	if err := json.Unmarshal(this.Ctx.Input.RequestBody, &orgRepo); err != nil {
-		this.Data["json"] = err.Error()
+		reason = err
+		statusCode = 400
 		return
 	}
 
+	cla := &models.CLA{ID: orgRepo.CLAID}
+	err := cla.Get()
+	if err != nil {
+		reason = fmt.Errorf("error finding the cla(id:%s), err: %v", cla.ID, err)
+		statusCode = 400
+		return
+	}
+
+	if cla.Language == "" {
+		reason = fmt.Errorf("the language of cla(id:%s) is empty", cla.ID)
+		statusCode = 500
+		return
+	}
+
+	orgRepo.CLALanguage = cla.Language
 	r, err := orgRepo.Create()
 	if err != nil {
-		this.Data["json"] = err.Error()
+		reason = err
+		statusCode = 500
 		return
 	}
 
@@ -42,21 +63,30 @@ func (this *OrgRepoController) Post() {
 // @Title Unbind CLA to Org/Repo
 // @Description unbind cla
 // @Param	uid		path 	string	true		"The uid of binding"
-// @Success 200 {string} delete success!
+// @Success 204 {string} delete success!
 // @Failure 403 uid is empty
 // @router /:uid [delete]
 func (this *OrgRepoController) Delete() {
+	var statusCode = 204
+	var reason error
+
 	defer func() {
-		this.ServeJSON()
+		sendResponse(&this.Controller, statusCode, reason)
 	}()
 
 	uid := this.GetString(":uid")
+	if uid == "" {
+		reason = fmt.Errorf("missing binding id")
+		statusCode = 400
+		return
+	}
 
 	orgRepo := models.OrgRepo{ID: uid}
 
 	err := orgRepo.Delete()
 	if err != nil {
-		this.Data["json"] = err.Error()
+		reason = err
+		statusCode = 500
 		return
 	}
